factories (#70)
* First sketch of factories implementation.

The idea here is to replace the ProtoLookupUtil class with something
that uses method calls instead of magic symbol lookup. In addition to
this, it'd be ideal if this was done statically insteaad of dynamically.
The reason for a preference over dynamic code (say method missing or
define method) is that then IDEs and other tooling can provide
autocompletes for users of the library.

So, the solution: it loads in all the protos under google ads (and then
some in the blast raadius) and after each one is defined, it stores
them in an array of known protos. Then it renders out an erb template
with that data.

Right now this is a hacky prototype but I was easily able to port a
campaign addition script over. Next I'll try somethinig meatier and
properly defining all resources/services.

* Require all protos

* Heavily refactor and clean up resource introspection code.

* Implement class mapping and filtering.

* Directory usage refactor

* Codegen refactor

* codegen cleanup

* Codegen example

* Continuing to hack on this.

* Add tests for tracepoints

* Add some tests, start to implement enums.

* Finish enums implementation, enables enumeration e.g.:

```
[1] pry(main)> require 'google/ads/google_ads/google_ads_client'
[2] pry(main)> gac = Google::Ads::GoogleAds::GoogleAdsClient.new
[3] pry(main)> gac.enum.ad_type.each { |x| p x }
:SHOPPING_PRODUCT_AD
:GMAIL_AD
:UNKNOWN
:UNSPECIFIED
:CALL_ONLY_AD
:VIDEO_AD
:IMAGE_AD
:EXPANDED_DYNAMIC_SEARCH_AD
:RESPONSIVE_DISPLAY_AD
:TEXT_AD
:LEGACY_RESPONSIVE_DISPLAY_AD
:LEGACY_APP_INSTALL_AD
:APP_AD
:SHOPPING_SMART_AD
:EXPANDED_TEXT_AD
:HOTEL_AD
:RESPONSIVE_SEARCH_AD
=> nil
[4] pry(main)>
```

* Make services work in factories implementation

* add standard for factory formatting

* Add a script to run codegen

* implement operations codegen

* Lint codegen

* Example of making updates all new

* fix update example

* Add examples of working with factories

* reset examples to master

* remove this example that isn't supposed to be here

* Network settings with factories!

* Add remove ops

* first pass of docs

* remove yarddoc

* fix gitignore for yard doc

* Fix filter name

* Add docstrings for updates

* its

* document remove

* documentation nit

* remove unneeded heredoc

* Make it work version independently

* doc nit

* Dynamically define proto modules for versions

* cleanup

* Add codegen to rakefile

* Remove operation wart from to level operation name

* Add a readme link to the factories docs

* Release 2.1.0