Fix codegen dependencies. (#188)
Previously I was running `bundle exec rake build` to make the gem usable
from scratch. However, this is not actually ideal for dev/testing.
Codegen needs patch and third party code, so this better expresses the
real dependencies that we have so that we can build for dev without
having to actually compile a gem.